Add files via upload
Tic tac Toe  game involves 2 players placing their respective symbols in a 3x3 grid. The player who manages to place three of their symbols in horizontal/vertical/diagonal row wins the game. If either player fails to do so the game ends in a draw.
This game is a very basic game as we all have played in our childhood days.